Reorganize the cli test as html.js
